requirements.txt
langchain
langchain-community
langchain-huggingface
langgraph
huggingface-hub
python-dotenv